+++ core/network/src/main/java/com/jparkbro/network/actor/ActorDataSource.kt
@@ -3,5 +3,5 @@ package com.jparkbro.network.actor
 import com.jparkbro.model.dto.info.InfoActorResponse
 
 interface ActorDataSource {
-    suspend fun getDetailActor(animeId: Int): Result<List<InfoActorResponse>>
+    suspend fun getDetailActor(animeId: Long): Result<List<InfoActorResponse>>
 }
